#R105.1 create the 'specialties list page' in admin account page
