+++ src/components/institutionOfEducationAdmin/ourSpecialties/ourSpecialties.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/index.scss';
+@import '../../../styles/index.scss';
 .specialityAdminPage {
   margin: 0 auto;
   width: calc((1180 / 1440) * 100%);
